+++ cpp/deque.hpp
@@ -18,7 +18,7 @@ Deque() {
   delete list;
 }
 
-T at(unsigned long long index) {
+T at(uint64_t index) {
   return list->at(index);
 }
 
@@ -50,7 +50,7 @@ T back() {
   return at(size() - 1);
 }
 
-unsigned long long size() {
+uint64_t size() {
   return list->length();
 }
